closes #6583