@MarcoGorelli 2023-02-06
🤔 this is noticeably slower than `master` https://www.kaggle.com/code/marcogorelli/polars-timing?scriptVersionId=118351982:

`master`:
- 3.77 s ± 39.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 3.74 s ± 59.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 3.73 s ± 13.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

here:
- 4.63 s ± 39.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 4.6 s ± 62.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 4.66 s ± 75.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)