@MarcoGorelli 2023-02-06
Right, performance is salvaged now, thanks for having pushed me on this, I learned a tonne! 🙏 

https://www.kaggle.com/code/marcogorelli/polars-timing/notebook?scriptVersionId=118393630

here:

*** naive ***
- 1.95 s ± 35.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 1.97 s ± 34.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 1.95 s ± 15.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
*** aware ***
- 3.33 s ± 47.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 3.32 s ± 34.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 3.32 s ± 31.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

master:

*** naive ***
- 1.98 s ± 31.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 1.96 s ± 21.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 1.97 s ± 39.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
*** aware ***
- 3.78 s ± 67 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 3.74 s ± 45.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- 3.76 s ± 73.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)