@ritchie46 2023-02-08
Great! I see you went with the outer branching approach. Yeap, that should be fastest (without doing crazy things). A function behind a reference can save some compiled code and also remove branches, but the indirection is often costly and it depends if its worth it.

In any case. Looks great!